refactor: create routes module
