+++ src/config/index.ts
@@ -1,0 +1,3 @@
+export default () => ({
+  port: parseInt(process.env.PORT, 10) || 4000,
+});
