+++ src/routes/album/album.interface.ts
