+++ src/routes/album/dto/create-album.dto.ts
@@ -1,26 +1,20 @@
 import {
-  IsDate,
   IsInt,
   IsNotEmpty,
-  IsPositive,
   IsString,
   IsUUID,
   ValidateIf,
 } from 'class-validator';
-import { UUID_VERSION } from '../../shared/const/uuid';
+import { UUID_VERSION } from '../../../shared/const/uuid';
 
-export class CreateTrackDto {
+export class CreateAlbumDto {
   @IsNotEmpty()
   @IsString()
   readonly name: string;
-  @IsPositive()
   @IsInt()
   @IsNotEmpty()
-  readonly duration: number;
+  readonly year: number;
   @IsUUID(UUID_VERSION)
   @ValidateIf((_, value) => value !== null)
   readonly artistId: string | null;
-  @IsUUID(UUID_VERSION)
-  @ValidateIf((_, value) => value !== null)
-  readonly albumId: string | null;
 }
