+++ src/routes/artist/artist.interface.ts
