+++ src/routes/artist/dto/create-artist.dto.ts
@@ -1,0 +1,10 @@
+import { IsBoolean, IsNotEmpty, IsString } from 'class-validator';
+
+export class CreateArtistDto {
+  @IsNotEmpty()
+  @IsString()
+  readonly name: string;
+  @IsBoolean()
+  @IsNotEmpty()
+  readonly grammy?: boolean;
+}
