+++ src/routes/favorites/favorites.interface.ts
