+++ src/routes/routes.module.ts
@@ -1,0 +1,24 @@
+import { Module } from '@nestjs/common';
+import { TrackModule } from './track/track.module';
+import { UserModule } from './user/user.module';
+import { ArtistModule } from './artist/artist.module';
+import { AlbumModule } from './album/album.module';
+import { FavoritesModule } from './favorites/favorites.module';
+
+@Module({
+  imports: [
+    UserModule,
+    TrackModule,
+    ArtistModule,
+    AlbumModule,
+    FavoritesModule,
+  ],
+  exports: [
+    UserModule,
+    TrackModule,
+    ArtistModule,
+    AlbumModule,
+    FavoritesModule,
+  ],
+})
+export class RoutesModule {}
