+++ src/routes/user/user.interface.ts
