+++ src/routes/user/user.module.ts
@@ -1,7 +1,6 @@
 import { Module } from '@nestjs/common';
 import { UserController } from './user.controller';
 import { UserService } from './user.service';
-import { PrismaModule } from '../prisma/prisma.module';
 
 @Module({
   controllers: [UserController],
